Put operators into module attributes
Arguably over engineered. However, I don't like to repeat the
"magic" string of all the operands. Feels wrong.

Arguably, the functions that handle the single operands could
be metaprogrammed entirely to also get rid of this duplication.
But, I don't want to start a meta programming exercise at this
point of the day and as long as we don't expect the list to grow
and to grow in a uniform way that'd arguably be even more
over engineering as meta programming should be a last resort.